requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-annoying==0.10.6
django-cors-headers==3.2.1
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
filelock==3.0.12
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
